Changes:
- move generate_proxy_model to Model
- add proxy mdoel shim with methods readded
- fix typings
